requirements.txt
asgiref==3.7.2
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.4
django-autoslug==1.9.9
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-flex-fields==1.0.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
environ==1.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
keyring==10.6.0
keyrings.alt==3.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pyasn1==0.5.0
pycparser==2.21
pycrypto==2.6.1
PyGObject==3.26.1
PyJWT==2.8.0
python-apt==1.6.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.25
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==2.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
